refactor: externalize SeedModal component
